move common error handling utilities from cuda_utils.h to error.hpp
